requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==1.10.3
attrs==24.2.0
beautifulsoup4==4.8.2
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
compressed_rtf==1.0.6
cryptography==43.0.3
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lightrag-hku==1.0.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.13.4
numpy==2.1.3
olefile==0.47
openai==1.54.4
packaging==24.2
pandas==2.2.3
pdfminer.six==20191110
pillow==11.0.0
protobuf==5.28.3
pyarrow==18.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-pptx==0.6.23
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.12.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wrapt==1.16.0
xlrd==1.2.0
XlsxWriter==3.2.0
